Extract priority logic into election package
