+++ cmd/pbm-agent/backup.go
@@ -8,6 +8,7 @@ import (
 	"github.com/percona/percona-backup-mongodb/pbm/config"
 	"github.com/percona/percona-backup-mongodb/pbm/ctrl"
 	"github.com/percona/percona-backup-mongodb/pbm/defs"
+	"github.com/percona/percona-backup-mongodb/pbm/election"
 	"github.com/percona/percona-backup-mongodb/pbm/errors"
 	"github.com/percona/percona-backup-mongodb/pbm/lock"
 	"github.com/percona/percona-backup-mongodb/pbm/log"
@@ -172,7 +173,7 @@ func (a *Agent) Backup(ctx context.Context, cmd *ctrl.BackupCmd, opid ctrl.OPID,
 			validCandidates = append(validCandidates, s)
 		}
 
-		nodes, err := BcpNodesPriority(ctx, a.leadConn, c, validCandidates)
+		nodes, err := election.BcpNodesPriority(ctx, a.leadConn, c, validCandidates)
 		if err != nil {
 			l.Error("get nodes priority: %v", err)
 			return
