+++ pbm/election/priority.go
@@ -1,4 +1,4 @@
-package main
+package election
 
 import (
 	"context"
